prival/requirements.txt
Flask==2.2.3
requests==2.28.1
psycopg2==2.9.6
python-dotenv==0.21.0
Django==4.2.6
djangorestframework==3.14.0